@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.11",
+    "@angular/animations": "^18.1.0",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.11",
-    "@angular/compiler": "^17.3.11",
-    "@angular/core": "^17.3.11",
-    "@angular/forms": "^17.3.11",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.11",
-    "@angular/platform-browser-dynamic": "^17.3.11",
-    "@angular/router": "^17.3.11",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/router": "^18.1.0",
     "bootstrap": "^4.6.2",
     "jquery": "^3.5.1",
     "rxjs": "^6.6.3",
@@ -28,9 +28,9 @@
     "zone.js": "~0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
-    "@angular/cli": "~17.3.8",
-    "@angular/compiler-cli": "^17.3.11",
+    "@angular-devkit/build-angular": "^18.1.0",
+    "@angular/cli": "~18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
     "@types/jasmine": "~4.0.0",
     "jasmine-core": "~4.3.0",
     "karma": "~6.4.0",
